+++ src/pages/Announcement/EditModal.tsx
@@ -56,7 +56,9 @@ export const EditModal: React.FC<EditModalProps> = (props) => {
               <Divider type="vertical" />
               <Typography.Text copyable={true}>{props.announcement.value?._id}</Typography.Text>
             </Form.Item>
-            <Form.Item label="发布于">{stringToYMD(props.announcement.value?.create_at)}</Form.Item>
+            <Form.Item label="发布于">
+              {stringToYMD(props.announcement.value?.create_at)}
+            </Form.Item>
             <Form.Item label="最后修改于">
               {stringToYMD(props.announcement.value?.update_at)}
             </Form.Item>