+++ src/pages/Dsiqus/Thread/index.tsx
@@ -9,7 +9,13 @@ import { useLoading } from 'veact-use'
 import classnames from 'classnames'
 import { Button, Card, Table, Select, Tag, Space, Switch, Divider, Typography } from 'antd'
 import * as Icon from '@ant-design/icons'
-import { getConfig, getThreads, ThreadState, OrderType, GeneralDisqusParams } from '@/store/disqus'
+import {
+  getConfig,
+  getThreads,
+  ThreadState,
+  OrderType,
+  GeneralDisqusParams,
+} from '@/store/disqus'
 import { stringToYMD } from '@/transforms/date'
 import { scrollTo } from '@/services/scroller'